+++ ouds_core/lib/components/pin_code_input/internal/modifier/ouds_pin_code_input_border_modifier.dart
@@ -15,8 +15,8 @@
 library;
 
 import 'package:flutter/material.dart';
-import 'package:ouds_theme_contract/ouds_theme.dart';
 import 'package:ouds_core/components/pin_code_input/internal/ouds_pin_code_input_control_state.dart';
+import 'package:ouds_theme_contract/ouds_theme.dart';
 
 /// A class that provides the border color for the OudsPinCodeInput based on its state and selection
 class OudsPinCodeInputBorderModifier {
@@ -26,29 +26,29 @@ class OudsPinCodeInputBorderModifier {
 
   /// Gets the borderSide based on the pin code input state and whether it is selected
   Border getPinCodeBorder(OudsPinCodeInputControlState state, [bool isError = false, bool? isOutlined]) {
-      switch (state) {
-        case OudsPinCodeInputControlState.enabled:
-          return Border(
-            bottom: getPinCodeBorderSideByState(state,isError),
-            top: !isOutlined! ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-            left: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-            right: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-          );
-        case OudsPinCodeInputControlState.hovered:
-          return Border(
-            bottom: getPinCodeBorderSideByState(state,isError),
-            top: !isOutlined! ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-            left: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-            right: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-          );
-        case OudsPinCodeInputControlState.focused:
-          return Border(
-            bottom: getPinCodeBorderSideByState(state,isError),
-            top: !isOutlined! ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-            left: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-            right: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state,isError),
-          );
-      }
+    switch (state) {
+      case OudsPinCodeInputControlState.enabled:
+        return Border(
+          bottom: getPinCodeBorderSideByState(state, isError),
+          top: !isOutlined! ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+          left: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+          right: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+        );
+      case OudsPinCodeInputControlState.hovered:
+        return Border(
+          bottom: getPinCodeBorderSideByState(state, isError),
+          top: !isOutlined! ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+          left: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+          right: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+        );
+      case OudsPinCodeInputControlState.focused:
+        return Border(
+          bottom: getPinCodeBorderSideByState(state, isError),
+          top: !isOutlined! ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+          left: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+          right: !isOutlined ? BorderSide.none : getPinCodeBorderSideByState(state, isError),
+        );
+    }
   }
 
   /// Returns a [BorderSide] based on the given [OudsPinCodeInputControlState].
@@ -60,13 +60,16 @@ class OudsPinCodeInputBorderModifier {
 
     switch (state) {
       case OudsPinCodeInputControlState.enabled:
-        return isError? BorderSide(color: colorToken.actionNegativeEnabled, width: textInputToken.borderWidthDefault)
+        return isError
+            ? BorderSide(color: colorToken.actionNegativeEnabled, width: textInputToken.borderWidthDefault)
             : BorderSide(color: textInputToken.colorBorderEnabled, width: textInputToken.borderWidthDefault);
       case OudsPinCodeInputControlState.hovered:
-        return isError? BorderSide(color: colorToken.actionNegativeEnabled, width: textInputToken.borderWidthDefault)
+        return isError
+            ? BorderSide(color: colorToken.actionNegativeEnabled, width: textInputToken.borderWidthDefault)
             : BorderSide(color: textInputToken.colorBorderHover, width: textInputToken.borderWidthDefault);
       case OudsPinCodeInputControlState.focused:
-        return isError? BorderSide(color: colorToken.actionNegativeEnabled, width: textInputToken.borderWidthDefault)
+        return isError
+            ? BorderSide(color: colorToken.actionNegativeEnabled, width: textInputToken.borderWidthDefault)
             : BorderSide(color: textInputToken.colorBorderFocus, width: textInputToken.borderWidthFocus);
     }
   }